Restructure the project from Python modules into a single package
