+++ monodepth2/datasets/kitti_dataset.py
@@ -11,7 +11,7 @@ import skimage.transform
 import numpy as np
 import PIL.Image as pil
 
-from kitti_utils import generate_depth_map
+from monodepth2.kitti_utils import generate_depth_map
 from .mono_dataset import MonoDataset
