+++ monodepth2/datasets/mono_dataset.py
@@ -6,10 +6,8 @@
 
 from __future__ import absolute_import, division, print_function
 
-import os
 import random
 import numpy as np
-import copy
 from PIL import Image  # using pillow-simd for increased speed
 
 import torch
